requirements.txt
numpy
Pillow  # provides PIL
scipy
google-cloud-storage
